requirements.txt
streamlit>=1.20.0
openai>=0.27.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0
requests>=2.28
scikit-learn>=1.2
mlflow>=2.4
tqdm>=4.65
pytest>=7.0
typing-extensions
sentence-transformers>=2.2.2
torch>=2.2.0
